Attendance working though lot of bugs there

+++ src/main/java/com/shehab/udms/DTO/CourseDTO.java
@@ -6,7 +6,7 @@ public record CourseDTO(
         Long id,
         String courseCode,
         String courseName,
-        int courseSemester,
+        Long courseSemester,
         double courseCredit,
         Department courseDepartment,
