+++ src/main/java/com/shehab/udms/DTO/StudentDTO.java
@@ -14,7 +14,7 @@ public record StudentDTO(
         String phone,
         String address,
         Department department,
-        Integer currSemester,
+        Long currSemester,
         Gender gender,
         Long userId,    // From nested Users
         Role role       // From nested Users
