+++ src/main/java/com/shehab/udms/repo/CourseRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CourseRepo extends JpaRepository<Course,Integer> {
+public interface CourseRepo extends JpaRepository<Course,Long> {
     Course findByCourseCode(String code);
 }
