+++ src/main/java/com/shehab/udms/repo/SemesterRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SemesterRepo extends JpaRepository<Semester, Integer> {
+public interface SemesterRepo extends JpaRepository<Semester, Long> {
 }
